requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
fastapi_mcp==0.1.6
alembic==1.13.1
pydantic==2.11.5 